refactor(#1115): move Narrator persona to world/scenes/narrator (re-export from combat)
